+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Upgraded the project to be compatible with `2019.1.3f1`.
 
+### Fixed
+
+- Fixed a bug where a worker's `World` could get disposed multiple times if you stopped the application inside the Editor while the worker is being created.
+
 ## `0.2.2` - 2019-05-15
 
 ### Breaking Changes